+++ bindingcollectionadapter/src/main/java/me/tatarka/bindingcollectionadapter/BindingCollectionAdapters.java
@@ -1,5 +1,6 @@
 package me.tatarka.bindingcollectionadapter;
 
+import android.content.Context;
 import android.content.res.Resources;
 import android.databinding.BindingAdapter;
 import android.databinding.ViewDataBinding;
@@ -8,6 +9,7 @@ import android.support.v4.view.ViewPager;
 import android.support.v7.widget.RecyclerView;
 import android.widget.ListView;
 
+import java.lang.reflect.Field;
 import java.util.Collection;
 
 /**
@@ -223,8 +225,40 @@ public class BindingCollectionAdapters {
      * Object)} returns false.
      */
     public static void throwMissingVariable(ViewDataBinding binding, int bindingVariable, @LayoutRes int layoutRes) {
-        Resources resources = binding.getRoot().getResources();
+        Context context = binding.getRoot().getContext();
+        Resources resources = context.getResources();
         String layoutName = resources.getResourceName(layoutRes);
-        throw new IllegalStateException("Could not bind variable '" + bindingVariable + "' in layout '" + layoutName + "'");
+        // Yeah reflection is slow, but this only happens when there is a programmer error.
+        String bindingVariableName;
+        try {
+            bindingVariableName = getBindingVariableName(context, bindingVariable);
+        } catch (Resources.NotFoundException e) {
+            // Fall back to int
+            bindingVariableName = "" + bindingVariable;
+        }
+        throw new IllegalStateException("Could not bind variable '" + bindingVariableName + "' in layout '" + layoutName + "'");
+    }
+
+    /**
+     * Returns the name for the given binding variable int. Warning! This uses reflection so it
+     * should <em>only</em> be used for debugging.
+     *
+     * @throws Resources.NotFoundException if the name cannot be found.
+     */
+    public static String getBindingVariableName(Context context, int bindingVariable) throws Resources.NotFoundException {
+        try {
+            String packageName = context.getApplicationInfo().packageName;
+            Class BRClass = Class.forName(packageName + ".BR");
+            Field[] fields = BRClass.getFields();
+            for (Field field : fields) {
+                int value = field.getInt(null);
+                if (value == bindingVariable) {
+                    return field.getName();
+                }
+            }
+        } catch (Exception e) {
+            //Ignore, throws NotFoundException
+        }
+        throw new Resources.NotFoundException("" + bindingVariable);
     }
 }